requirements.txt
fastapi
uvicorn
python-multipart
pymongo[srv]
requests
speechrecognition
python-dotenv
